requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
websockets>=12.0
jinja2>=3.1.3

# Local LLM (Ollama client)
ollama>=0.4.0

# Cloud LLM providers
anthropic>=0.42.0
openai>=1.60.0
google-genai>=1.0.0

# Speech-to-text
faster-whisper>=1.1.0

# TTS
piper-tts>=1.4.1

# Audio processing
numpy>=1.26.0
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
Pillow>=10.3.0
pydantic>=2.7.0
